Add memory allocation check and function renaming
In saliency_map.c file, add memory allocation failure check and
remove redundant memory allocation. Also done some function/param
renaming and data type change.

NO_STATS_CHANGE

+++ av1/encoder/encoder.h
@@ -3484,7 +3484,7 @@ typedef struct AV1_COMP {
   /*!
    * Pixel level saliency map for each frame.
    */
-  double *saliency_map;
+  uint8_t *saliency_map;
 } AV1_COMP;
 
 /*!
